Updating app requirements for Django 1.8 support.
The Django version in the requirements has been upgraded to version 1.8.
django-cms>3.1.1 is now required for Django 1.8 support. A fork of
django-shop is also required for Django 1.8 support.

The fork is installed with an egg name of "django-shop" so that the
dependency on "django-shop" within setup.py doesn't pull in the
non-forked release from pypy.